# backend/institutotcnicorical-1ec-Pergola-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/institutotcnicorical-1ec/Pergola/1.0.0"
  } ],
  "info" : {
    "description" : "EL API de Pérgola se encarga de manejar la base de datos de la empresa de Joyería Pérgola.",
    "version" : "1.0.0",
    "title" : "Pergola API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/api/employees" : {
      "get" : {
        "summary" : "GET EMPLOYEES",
        "tags" : [ "Employees" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "get-api-employees",
        "description" : "Listar empleados."
      },
      "post" : {
        "summary" : "POST EOMPLOYEES",
        "tags" : [ "Employees" ],
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "operationId" : "post-api-employees",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "name", "lastName", "username", "email", "phoneNumber", "birthDate", "DUI", "password", "userType", "hireDate", "isVerified" ],
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "br1v24nwgvui1"
                    }
                  },
                  "lastName" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "wdm53tvl1uf0f"
                    }
                  },
                  "username" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "nuym4f0flfayv"
                    }
                  },
                  "email" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "4j8cf8s3cesph"
                    },
                    "format" : "email"
                  },
                  "phoneNumber" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "08oa64mlkfe3q"
                    }
                  },
                  "birthDate" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "45or6p66pz0tl"
                    },
                    "format" : "date"
                  },
                  "DUI" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "qxgfnlqzj30wi"
                    }
                  },
                  "password" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "pmy3f5r82j4sj"
                    },
                    "format" : "password"
                  },
                  "userType" : {
                    "x-stoplight" : {
                      "id" : "x5xbicmuo0dzb"
                    },
                    "enum" : [ "Administrador", "Empleado" ]
                  },
                  "hireDate" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "8p4dyx0plwbvh"
                    },
                    "format" : "date"
                  },
                  "isVerified" : {
                    "type" : "boolean",
                    "x-stoplight" : {
                      "id" : "5sjhk1t031h0u"
                    }
                  }
                }
              }
            }
          }
        },
        "description" : "Registrar un nuevo empleado (con foto)."
      }
    },
    "/employees/:id" : {
      "get" : {
        "summary" : "GET EMPLOYEES - :ID",
        "tags" : [ "Employees" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "get-employees-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "description" : "Obtener datos de un empleado específico"
      },
      "put" : {
        "summary" : "PUT EMPLOYEES",
        "tags" : [ "Employees" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "operationId" : "put-employees-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "name", "lastName", "username", "email", "phoneNumber", "birthDate", "DUI", "password", "userType", "hireDate", "isVerified" ],
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "wun7ecnynyszi"
                    }
                  },
                  "lastName" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "k9jle3691kdjp"
                    }
                  },
                  "username" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "esdr4e39x6d2i"
                    }
                  },
                  "email" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "i10rzdnet8p2c"
                    },
                    "format" : "email"
                  },
                  "phoneNumber" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "41w9m314qqft1"
                    }
                  },
                  "birthDate" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "g2idifqz1p2ld"
                    },
                    "format" : "date"
                  },
                  "DUI" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "mf7bv76nezwi2"
                    }
                  },
                  "password" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "7ecqqq7oaqn2d"
                    },
                    "format" : "password"
                  },
                  "userType" : {
                    "x-stoplight" : {
                      "id" : "umzgpdvojipj9"
                    },
                    "enum" : [ "admin", "colaborador" ]
                  },
                  "hireDate" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "whcxqqp5gyeb1"
                    },
                    "format" : "date"
                  },
                  "isVerified" : {
                    "type" : "boolean",
                    "x-stoplight" : {
                      "id" : "3a5mxujrb6bxg"
                    }
                  }
                }
              }
            }
          }
        },
        "description" : "Actualizar datos de un empleado (con foto)."
      },
      "delete" : {
        "summary" : "DELETE EMPLOYEES",
        "tags" : [ "Employees" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "delete-employees-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "description" : "Eliminar un empleado."
      }
    },
...
  },
...
}